Added resource and converter
Add resource and corresponding impl class which uses the generated DTO as request or response type (though return type is still jaxrs response. But basically the idea is to use the generated dto classes for all ws transactions.
Corresponding test is added as well which includes basic mockito setup

Converter classes is also in place. This should handle the conversion from domain to dto and vice versa.
Spring's coversion services will be used